I created a basic project.
The system so far has only one machine.  This machine must be started using docker rather than docker-compose because of unknown reasons.

The machine is a decent fedora 23 customized machine.  It has a workspace, prompt, gpg, ssh, man, etc.

Note that for development purposes:
There are two primary times:
* install - things that happen before the image is built
* init - things that happen after the user executes run

The more things that can happen in install the better.

install time is divided into layers.  Layers are cached.  during build, no layer is rebuilt if the cached layer can be used.

The biggest layer can take 10-20 minutes to build.  if you change something and it invalidates that layer, then the build will take an addition 10-20 minutes.

Therefore for development,
 * for things needed to happen at build time:  try to add a subsequent layer rather than modify the 10-20 minute layer.
 * for things needed to happen at init time:  try to add a script to init/{s,bin}.  it will be executed after the normal init and it won't invalidate the 10-20 minute layer.

After you have built and tested, increment the minor version, tag, and migrate the improvements into the 10-20 minute layer (because fewer layers are a best practice).